# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_auth.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/internal_auth.User'
    type: object
  internal_auth.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_auth.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
  internal_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_auth.MessageResponse:
    properties:
      message:
        type: string
    type: object
  internal_auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_auth.RegisterResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/internal_auth.User'
    type: object
  internal_auth.ResendVerificationRequest:
    properties:
      email:
        type: string
    type: object
  internal_auth.ResetPasswordRequest:
    properties:
      newPassword:
        type: string
      token:
        type: string
    type: object
  internal_auth.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      emailVerified:
        type: boolean
      id:
        type: string
      updatedAt:
        type: string
    type: object
  internal_auth.VerifyEmailRequest:
    properties:
      token:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/trackable/trackable
  description: KPI tracking platform API
  license:
    name: MIT
  title: Trackable API
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset email to user
      parameters:
      - description: User email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Request password reset
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      description: Logout the current user (client should discard token)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.MessageResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/me:
    get:
      description: Get the currently authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_auth.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Resend email verification link to user
      parameters:
      - description: User email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Resend verification email
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password with token
      parameters:
      - description: Reset token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Reset password
      tags:
      - auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verify user email address with token
      parameters:
      - description: Verification token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Verify email
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"